backend/requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Google Gemini
google-generativeai==0.8.4

# Pinecone vector DB
pinecone-client==5.0.1

# Document processing
PyMuPDF==1.25.3
python-pptx==1.0.2
opencv-python-headless==4.10.0.84
Pillow==11.1.0
pytesseract==0.3.13

# PubMed / Bio
biopython==1.85
httpx==0.28.1

# Report generation
reportlab==4.2.5

# SSE streaming
sse-starlette==2.2.1

# Utilities
aiofiles==24.1.0
